Move vue-form tests
