+++ packages/vue-form/tests/useForm.test.tsx
@@ -1,11 +1,11 @@
 import { describe, expect, it, vi } from 'vitest'
-import userEvent from '@testing-library/user-event'
+import { userEvent } from '@testing-library/user-event'
 import { render, waitFor } from '@testing-library/vue'
 import { defineComponent, h, ref } from 'vue'
-import { useForm } from '../index'
+import { useForm } from '../src/index'
 import { sleep } from './utils'
 
-import type { FieldApi, ValidationErrorMap } from '../index'
+import type { FieldApi, ValidationErrorMap } from '../src/index'
 
 const user = userEvent.setup()
