+++ packages/vue-form/tsconfig.json
@@ -9,11 +9,5 @@
       "@tanstack/form-core": ["../form-core/src"]
     }
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    ".eslintrc.cjs",
-    "test-setup.ts",
-    "vite.config.ts"
-  ]
+  "include": ["src", "tests", ".eslintrc.cjs", "vite.config.ts"]
 }
